Use 32bit address for mapped addresses

+++ include/nesemu/mapper/internal/mapper_000.h
@@ -9,7 +9,7 @@ class Mapper000 : public Mapper {
 public:
   using Mapper::Mapper;  // "Inherit" constructor
 
-  uint16_t decodeCPUAddress(uint16_t addr) const override { return addr & (prg_banks_ > 1 ? 0x7FFF : 0x3FFF); };
+  uint32_t decodeCPUAddress(uint16_t addr) const override { return addr & (prg_banks_ > 1 ? 0x7FFF : 0x3FFF); };
 };
 
 }  // namespace mapper::internal
